Add function to send signal to child process race-free.
There is potentially a race between sending a signal to a child process
and that same process being reaped in another thread. In the worst (but
unlikely) case the process ID could be re-used by another process in the
system, leading to the signal being sent to the wrong process. So, add a
function to signal children in a race-free manner.