Make kqueue force emulation of reading regular files.
Default kqueue semantics are that a file is not "ready for reading" if
at EOF. This differs from select/poll and has limited use; in particular
it prevents adding a descriptor without knowing what kind it is.

On FreeBSD we can use NOTE_FILE_POLL to get the right semantics. For
systems without NOTE_FILE_POLL we instead check for regular files using
stat (which is hopefully not too expensive) and fall back to emulation.