Add a sigmaskf function to the loop traits for signal masking.
POSIX says that sigprocmask cannot be relied upon in multi-threaded
programs, but pthread_sigmask requires linking against pthreads.
Therefore we need to use one or the other depending on the situation.
For now, we use the mutex type to decide which (i.e. we use sigprocmask
only for null_mutex).