svector: fix to avoid constructing reserved-but-not-allocated elements.
svector with element type T was using a T[] backing store. The issue is
that allocating a T[] with some capacity c is going to construct c x T
objets. We want to avoid this construction and only construct elements
when they are actually added to the vector. To do this, use a union type
to obscure the underlying element.

(Note the same problem occurs with destruction, and that is also fixed
by this approach).