Fix (I think) quoting in cmake script generation
Cmake documentation isn't super clear, but it seems that splitting is
done before variable expansion (and unquoted variable references will
expand to an already-split list).

So for things like DASYNQ_INCLUDE_DIRS we don't want to quote it
generally (it's technically a list). Quoting it would force it to be
interpreted as a single value (list represented by a single string, ';'
delimited).