Fix pselect/select backend compilation
The file descriptor set copies that are passed to (p)select can't be
volatile (it's not possible to assign a volatile set from a non-volatile
set). Anyway they didn't need to be, since they were not modified
between setjmp/longjmp calls.

For pselect, the signal mask was a local variable which was used and
modified after setjmp, which is problematic; to fix that, move it out to
a member variable.

In addition, use an explicit signal fence (release) before unmasking
signals.