Re-work loop traits to be composable
As loop backend components are composable, so should be the traits.
Before, the traits class for the main backend component specified all
the traits, not those that are really the domain of subcomponents.

Now, traits are composable, and the `loop_t' template (used to
instantiate the event loop by composing with the event dispatch) is
provided by the composed traits object rather than needing to be
composed "manually" as a separate step.

As part of this, some backend mechanisms had their `fd_s.get_fd()`
member fixed (it was in several misspelt as `getFd()`).